refactor(flux): drop the v2 segment from Flux namespaces
Rename the Flux (v2 listen / STT) client and model namespaces to drop the
redundant `v2` segment, which collided in meaning with the SDK-architecture
`v2` (e.g. Speak.v2 targets /v1/speak). "Flux" is a v2-API-only product, so
the segment disambiguated nothing.

  Deepgram.Clients.Flux.v2.WebSocket -> Deepgram.Clients.Flux.WebSocket
  Deepgram.Models.Flux.v2.WebSocket  -> Deepgram.Models.Flux.WebSocket

Interfaces (Interfaces.v2.IFluxWebSocketClient), the FluxWebSocketClient
wrapper, and ClientFactory.CreateFluxWebSocketClient() are unchanged. Purely
mechanical: no behavior change; full suite green (211 passed, 1 live skipped).

Coordinated with the incoming Flux TTS (/v2/speak) work, which lands under
Deepgram.Clients.Flux.Speak.*.
