+++ Deepgram.Tests/UnitTests/ClientTests/FluxClientTests.cs
@@ -7,10 +7,10 @@ using System.Net.WebSockets;
 
 using Deepgram.Models.Authenticate.v1;
 using Deepgram.Models.Exceptions.v1;
-using Deepgram.Models.Flux.v2.WebSocket;
-using Deepgram.Clients.Flux.v2.WebSocket;
+using Deepgram.Models.Flux.WebSocket;
+using Deepgram.Clients.Flux.WebSocket;
 using Common = Deepgram.Models.Common.v2.WebSocket;
-using FluxConstants = Deepgram.Clients.Flux.v2.WebSocket.Constants;
+using FluxConstants = Deepgram.Clients.Flux.WebSocket.Constants;
 
 namespace Deepgram.Tests.UnitTests.ClientTests;
