+++ Deepgram/Clients/Flux/WebSocket/UriSegments.cs
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a MIT license that can be found in the LICENSE file.
 // SPDX-License-Identifier: MIT
 
-namespace Deepgram.Clients.Flux.v2.WebSocket;
+namespace Deepgram.Clients.Flux.WebSocket;
 
 /// <summary>
 /// (PREVIEW) URI segments for the Flux (v2 listen) WebSocket client.
