--- Deepgram/Models/Flux/v2/WebSocket/ErrorResponse.cs
@@ -1,51 +1,0 @@
-// Copyright 2026 Deepgram .NET SDK contributors. All Rights Reserved.
-// Use of this source code is governed by a MIT license that can be found in the LICENSE file.
-// SPDX-License-Identifier: MIT
-
-namespace Deepgram.Models.Flux.v2.WebSocket;
-
-/// <summary>
-/// (PREVIEW) Fatal error sent by the Deepgram Flux (v2 listen) API. The wire value of the
-/// "type" property is literally "Error" (the AsyncAPI schema names this message FatalError).
-/// The server closes the connection after sending it.
-/// </summary>
-public record ErrorResponse
-{
-    /// <summary>
-    /// Error event type. The wire value is "Error".
-    /// </summary>
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    [JsonPropertyName("type")]
-    [JsonConverter(typeof(JsonStringEnumConverter))]
-    public FluxType? Type { get; set; } = FluxType.Error;
-
-    /// <summary>
-    /// Starts at 0 and increments for each message the server sends to the client,
-    /// including messages of other types.
-    /// </summary>
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    [JsonPropertyName("sequence_id")]
-    public int? SequenceId { get; set; }
-
-    /// <summary>
-    /// A string code describing the error, e.g. INTERNAL_SERVER_ERROR.
-    /// </summary>
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    [JsonPropertyName("code")]
-    public string? Code { get; set; }
-
-    /// <summary>
-    /// Prose description of the error.
-    /// </summary>
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    [JsonPropertyName("description")]
-    public string? Description { get; set; }
-
-    /// <summary>
-    /// Override ToString method to serialize the object
-    /// </summary>
-    public override string ToString()
-    {
-        return Regex.Unescape(JsonSerializer.Serialize(this, JsonSerializeOptions.DefaultOptions));
-    }
-}
